Post-demo changes
