+++ driver-service/src/main/java/com/example/driverservice/dto/EntityRequest.java
@@ -1,0 +1,4 @@
+package com.example.driverservice.dto;
+
+public interface EntityRequest {
+}
