+++ driver-service/src/main/resources/db/changelog/v.1.0.0/changelog-v.1.0.0-accumulative-test.xml
@@ -10,6 +10,5 @@
 
     <include file="2025-01-28-create-tables.xml" relativeToChangelogFile="true"/>
     <include file="2025-02-19-add-rate-column.xml" relativeToChangelogFile="true"/>
-    <include file="2025-03-06-insert-initial-data-test.xml" relativeToChangelogFile="true"/>
 
 </databaseChangeLog>
