--- driver-service/src/test/java/com/example/driverservice/DriverServiceApplicationTests.java
@@ -1,13 +1,0 @@
-package com.example.driverservice;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class DriverServiceApplicationTests {
-
-    @Test
-    void contextLoads() {
-    }
-
-}
