+++ driver-service/src/test/java/com/example/driverservice/configuration/CucumberSpringConfiguration.java
@@ -1,0 +1,12 @@
+package com.example.driverservice.configuration;
+
+import com.example.driverservice.configuration.container.MySqlTestContainer;
+import io.cucumber.spring.CucumberContextConfiguration;
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.boot.testcontainers.context.ImportTestcontainers;
+
+@CucumberContextConfiguration
+@SpringBootTest(classes = TestConfig.class, webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
+@ImportTestcontainers(MySqlTestContainer.class)
+public class CucumberSpringConfiguration {
+}
