+++ driver-service/src/test/java/com/example/driverservice/configuration/TestConfig.java
@@ -1,0 +1,11 @@
+package com.example.driverservice.configuration;
+
+import org.springframework.boot.autoconfigure.EnableAutoConfiguration;
+import org.springframework.context.annotation.ComponentScan;
+import org.springframework.context.annotation.Configuration;
+
+@Configuration
+@ComponentScan(basePackages = "com.example.driverservice")
+@EnableAutoConfiguration
+public class TestConfig {
+}
