+++ driver-service/src/test/java/com/example/driverservice/configuration/container/MySqlTestContainer.java
@@ -1,0 +1,21 @@
+package com.example.driverservice.configuration.container;
+
+import org.springframework.boot.testcontainers.service.connection.ServiceConnection;
+import org.springframework.test.context.ActiveProfiles;
+import org.testcontainers.containers.MySQLContainer;
+import org.testcontainers.utility.DockerImageName;
+
+import static com.example.driverservice.configuration.constants.GeneralUtilityConstants.DB_DOCKER_IMAGE;
+
+@ActiveProfiles("test")
+public abstract class MySqlTestContainer {
+
+    @ServiceConnection
+    protected static final MySQLContainer<?> MYSQL_CONTAINER = new MySQLContainer<>(
+            DockerImageName.parse(DB_DOCKER_IMAGE));
+
+    static {
+        MYSQL_CONTAINER.start();
+    }
+
+}
