+++ driver-service/src/test/java/com/example/driverservice/e2e/DriverStepDefinitions.java
@@ -1,31 +1,67 @@
 package com.example.driverservice.e2e;
 
-import static com.example.driverservice.utility.constants.GeneralUtilityConstants.DRIVER_CONTROLLER_BASE_URI_DETERMINED;
-import static com.example.driverservice.utility.constants.GeneralUtilityConstants.ENDPOINT_WITH_ID;
-import static com.example.driverservice.utility.constants.GeneralUtilityConstants.ID_PARAMETER_NAME;
+import static com.example.driverservice.configuration.constants.GeneralUtilityConstants.CONTROLLER_BASE_URI;
+import static com.example.driverservice.configuration.constants.GeneralUtilityConstants.DRIVER_ENDPOINT;
+import static com.example.driverservice.configuration.constants.GeneralUtilityConstants.DRIVER_ENDPOINT_WITH_ID;
+import static com.example.driverservice.configuration.constants.GeneralUtilityConstants.ID_PARAMETER_NAME;
+import static com.example.driverservice.configuration.constants.SqlConstants.SQL_DELETE_ALL_TEST_DATA_CARS;
+import static com.example.driverservice.configuration.constants.SqlConstants.SQL_DELETE_ALL_TEST_DATA_DRIVERS;
+import static com.example.driverservice.configuration.constants.SqlConstants.SQL_INSERT_TEST_ENTITY_DATA_CARS;
+import static com.example.driverservice.configuration.constants.SqlConstants.SQL_INSERT_TEST_ENTITY_DATA_CARS_2;
+import static com.example.driverservice.configuration.constants.SqlConstants.SQL_INSERT_TEST_ENTITY_DATA_DRIVERS;
+import static com.example.driverservice.configuration.constants.SqlConstants.SQL_INSERT_TEST_ENTITY_DATA_DRIVERS_2;
 import static io.restassured.RestAssured.baseURI;
 import static io.restassured.RestAssured.given;
+import static io.restassured.RestAssured.port;
 import static org.assertj.core.api.Assertions.assertThat;
 
 import com.example.driverservice.dto.driver.DriverRequest;
 import com.example.driverservice.dto.driver.DriverResponse;
 import com.fasterxml.jackson.databind.ObjectMapper;
+import io.cucumber.java.Before;
 import io.cucumber.java.en.And;
 import io.cucumber.java.en.Given;
 import io.cucumber.java.en.Then;
 import io.cucumber.java.en.When;
 import io.restassured.http.ContentType;
 import io.restassured.response.Response;
+import jakarta.annotation.PostConstruct;
+import lombok.RequiredArgsConstructor;
 import lombok.SneakyThrows;
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.boot.test.web.server.LocalServerPort;
 import org.springframework.http.HttpStatus;
+import org.springframework.jdbc.core.JdbcTemplate;
 
+@SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
+@RequiredArgsConstructor
 public class DriverStepDefinitions {
 
-    private final ObjectMapper objectMapper = new ObjectMapper();
+    private final JdbcTemplate jdbcTemplate;
+    private final ObjectMapper objectMapper;
     private Long driverId;
     private Response response;
     private DriverRequest driverRequest;
     private DriverResponse actualResponse;
+
+    @LocalServerPort
+    private int curPort;
+
+    @PostConstruct
+    public void setUp() {
+        baseURI = CONTROLLER_BASE_URI;
+        port = curPort;
+    }
+
+    @Before
+    public void restoreDbData() {
+        jdbcTemplate.execute(SQL_DELETE_ALL_TEST_DATA_CARS);
+        jdbcTemplate.execute(SQL_DELETE_ALL_TEST_DATA_DRIVERS);
+        jdbcTemplate.execute(SQL_INSERT_TEST_ENTITY_DATA_DRIVERS);
+        jdbcTemplate.execute(SQL_INSERT_TEST_ENTITY_DATA_DRIVERS_2);
+        jdbcTemplate.execute(SQL_INSERT_TEST_ENTITY_DATA_CARS);
+        jdbcTemplate.execute(SQL_INSERT_TEST_ENTITY_DATA_CARS_2);
+    }
 
     @Given("driver with id {long}")
     public void findDriverById(Long id) {
@@ -47,49 +83,44 @@ public class DriverStepDefinitions {
 
     @When("searching driver by id")
     public void searchDriverById() {
-        baseURI = DRIVER_CONTROLLER_BASE_URI_DETERMINED;
         response = given()
                 .pathParam(ID_PARAMETER_NAME, driverId)
                 .when()
-                .get(ENDPOINT_WITH_ID);
+                .get(DRIVER_ENDPOINT_WITH_ID);
     }
 
     @When("getting all drivers")
     public void getAllDrivers() {
-        baseURI = DRIVER_CONTROLLER_BASE_URI_DETERMINED;
         response = given()
                 .when()
-                .get();
+                .get(DRIVER_ENDPOINT);
     }
 
     @When("saving driver")
     public void driverGetSaved() {
-        baseURI = DRIVER_CONTROLLER_BASE_URI_DETERMINED;
         response = given()
                 .contentType(ContentType.JSON)
                 .body(driverRequest)
                 .when()
-                .post();
+                .post(DRIVER_ENDPOINT);
     }
 
     @When("updating driver")
     public void driverUpdated() {
-        baseURI = DRIVER_CONTROLLER_BASE_URI_DETERMINED;
         response = given()
                 .pathParam(ID_PARAMETER_NAME, driverId)
                 .contentType(ContentType.JSON)
                 .body(driverRequest)
                 .when()
-                .patch(ENDPOINT_WITH_ID);
+                .patch(DRIVER_ENDPOINT_WITH_ID);
     }
 
     @When("deleting driver")
     public void driverDeleted() {
-        baseURI = DRIVER_CONTROLLER_BASE_URI_DETERMINED;
         response = given()
                 .pathParam(ID_PARAMETER_NAME, driverId)
                 .when()
-                .delete(ENDPOINT_WITH_ID);
+                .delete(DRIVER_ENDPOINT_WITH_ID);
     }
 
     @Then("received driver response has status {int}")
@@ -102,24 +133,12 @@ public class DriverStepDefinitions {
     @SneakyThrows
     public void driverBodyEqualsTo(String expectedDriverJson) {
         DriverResponse expectedResponse = objectMapper.readValue(expectedDriverJson, DriverResponse.class);
-        DriverResponse actualResponse = response.as(DriverResponse.class);
+        actualResponse = response.as(DriverResponse.class);
 
-        assertThat(expectedResponse)
-                .usingRecursiveComparison()
-                .ignoringFields(ID_PARAMETER_NAME)
-                .isEqualTo(actualResponse);
-    }
-
-    @And("driver body no carsId equals to")
-    @SneakyThrows
-    public void driverNoCarsBodyEqualsTo(String expectedDriverJson) {
-        DriverResponse expectedResponse = objectMapper.readValue(expectedDriverJson, DriverResponse.class);
-        DriverResponse actualResponse = response.as(DriverResponse.class);
-
-        assertThat(expectedResponse)
+        assertThat(actualResponse)
                 .usingRecursiveComparison()
                 .ignoringFields("carIds")
-                .isEqualTo(actualResponse);
+                .isEqualTo(expectedResponse);
     }
 
     @And("body after driver create operation equals to")
@@ -128,21 +147,30 @@ public class DriverStepDefinitions {
         DriverResponse expectedResponse = objectMapper.readValue(expectedDriverJson, DriverResponse.class);
         actualResponse = response.as(DriverResponse.class);
 
-        assertThat(expectedResponse)
+        assertThat(actualResponse)
                 .usingRecursiveComparison()
                 .ignoringFields(ID_PARAMETER_NAME)
-                .isEqualTo(actualResponse);
+                .isEqualTo(expectedResponse);
     }
 
-    @And("check if driver created in db")
+    @And("check if driver is created in db")
     public void checkIfDriverCreatedInDb() {
-        baseURI = DRIVER_CONTROLLER_BASE_URI_DETERMINED;
         given()
                 .pathParam(ID_PARAMETER_NAME, actualResponse.id())
                 .when()
-                .get(ENDPOINT_WITH_ID)
+                .get(DRIVER_ENDPOINT_WITH_ID)
                 .then()
                 .statusCode(HttpStatus.OK.value());
+    }
+
+    @And("check if driver is deleted in db")
+    public void checkIfDeletedInDb() {
+        given()
+                .pathParam(ID_PARAMETER_NAME, driverId)
+                .when()
+                .get(DRIVER_ENDPOINT_WITH_ID)
+                .then()
+                .statusCode(HttpStatus.NOT_FOUND.value());
     }
 
 }
