+++ driver-service/src/test/resources/sql/delete_all_test_data_cars.sql
@@ -1,0 +1,1 @@
+delete from cars;
