+++ driver-service/src/test/resources/sql/init_test_data_cars.sql
@@ -1,0 +1,5 @@
+insert into cars (id, brand, number, color, driver_id)
+values (1, 'Toyota', '1234AB-3', 'Red', 1);
+
+insert into cars (id, brand, number, color, driver_id)
+values (2, 'Ford', '1234AB-5', 'Blue', 2);
