+++ driver-service/src/test/resources/sql/init_test_data_drivers.sql
@@ -1,0 +1,5 @@
+insert into drivers (id, name, email, phone, sex, rate)
+values (1, 'John Cena', 'john@gmail.com', '+375291111111', 1, 4.52);
+
+insert into drivers (id, name, email, phone, sex, rate)
+values (2, 'Leeroy Jenkins', 'leeroy228@gmail.com', '+375292222222', 1, 4.73);
